fix missing updates to vue store when sorting answers
when the SET_ANSWERS alone is done, the answers appear to "jump back",
and it gets really weird, when also toggling the "correct"
checkbox. It's not the actual CAUSE of the weird behaviour though.
this change restores the function that updateAnswersOrders actually
changes the data in the store again, when the sort order of answers
changes in ui.

this also is a less visible problem with the live version, that the vue
store is only updated, when the checkmarks are updated or maybe the
content is edited. changing the order of answers alone doesn't trigger
an update, but the change remains in the ui, since no redraw is
triggered (see next change).

this is due to the fact, that vue apparently doesn't generally
rerender the elements, if the key is the same. the second change fixes
that and uses as key the actual id of the answer. now vue triggers
the updates for sorting order (no visual reverting).

(disclaimer: I was relying on the store always being in sync with the
display and noticed stuff not changing in the store when re-sorting
answer options.)